+++ lib/drive/google.js
@@ -1,14 +1,13 @@
-/* global */
-const {CustomError} = require("./error");
+/* global self */
+const {createRequest, RequestError} = require("../request");
 
 function createDrive({
   getAccessToken,
-  /* eslint-disable no-undef */
-  fetch: _fetch = fetch,
-  FormData: _FormData = FormData,
-  Blob: _Blob = Blob
-  /* eslint-enable */
+  fetch = (typeof self !== "undefined" ? self : global).fetch,
+  FormData = (typeof self !== "undefined" ? self : global).FormData,
+  Blob = (typeof self !== "undefined" ? self : global).Blob
 }) {
+  const request = createRequest({fetch, getAccessToken});
   const fileMetaCache = new Map;
   let lockRev;
   return {
@@ -25,17 +24,16 @@ function createDrive({
   };
   
   async function revDelete(fileId, revId) {
-    await query({
+    await request({
       method: "DELETE",
-      path: `https://www.googleapis.com/drive/v3/files/${fileId}/revisions/${revId}`,
-      format: null
+      path: `https://www.googleapis.com/drive/v3/files/${fileId}/revisions/${revId}`
     });
   }
   
   async function acquireLock(expire) {
     const lock = fileMetaCache.get("lock.json");
     const {headRevisionId} = await queryPatch(lock.id, JSON.stringify({expire: Date.now() + expire * 60 * 1000}));
-    const result = await query({
+    const result = await request({
       path: `https://www.googleapis.com/drive/v3/files/${lock.id}/revisions?fields=revisions(id)`
     });
     for (let i = 1; i < result.revisions.length; i++) {
@@ -45,13 +43,13 @@ function createDrive({
         lockRev = headRevisionId;
         return;
       }
-      const rev = await query({
+      const rev = await request({
         path: `https://www.googleapis.com/drive/v3/files/${lock.id}/revisions/${revId}?alt=media`
       });
       if (rev.expire > Date.now()) {
         // failed, delete the lock
         await revDelete(lock.id, headRevisionId);
-        throw new CustomError("EEXIST", new Error("failed to acquire lock"));
+        throw new RequestError("failed to acquire lock", null, "EEXIST");
       }
       // delete outdated lock
       await revDelete(lock.id, revId);
@@ -67,16 +65,16 @@ function createDrive({
   
   async function queryList(path, onPage) {
     path = "https://www.googleapis.com/drive/v3/files?spaces=appDataFolder&fields=nextPageToken,files(id,name,headRevisionId)" + (path ? "&" + path : "");
-    let result = await query({path});
+    let result = await request({path});
     onPage(result);
     while (result.nextPageToken) {
-      result = await query({path: `${path}&pageToken=${result.nextPageToken}`});
+      result = await request({path: `${path}&pageToken=${result.nextPageToken}`});
       onPage(result);
     }
   }
   
   async function queryPatch(id, text) {
-    return await query({
+    return await request({
       method: "PATCH",
       path: `https://www.googleapis.com/upload/drive/v3/files/${id}?uploadType=media&fields=headRevisionId`,
       headers: {
@@ -107,24 +105,6 @@ function createDrive({
     }
   }
   
-  async function query({method = "GET", path, headers, format = "json", ...args}) {
-    const res = await _fetch(path, {
-      method,
-      headers: {
-        "Authorization": `Bearer ${await getAccessToken()}`,
-        ...headers
-      },
-      ...args
-    });
-    if (!res.ok) {
-      const {error} = await res.json();
-      throw new CustomError(error.code, error);
-    }
-    if (format) {
-      return await res[format]();
-    }
-  }
-  
   async function list(file) {
     // FIXME: this only works if file is a single dir
     // FIXME: this only works if the list method is called right after init, use
@@ -140,13 +120,12 @@ function createDrive({
       await updateMeta(`name = '${file}'`);
       meta = fileMetaCache.get(file);
       if (!meta) {
-        throw new CustomError("ENOENT", new Error(`metaCache doesn't contain ${file}`));
+        throw new RequestError(`metaCache doesn't contain ${file}`, null, "ENOENT");
       }
     }
     try {
-      return await query({
-        path: `https://www.googleapis.com/drive/v3/files/${meta.id}?alt=media`,
-        format: "text"
+      return await request({
+        path: `https://www.googleapis.com/drive/v3/files/${meta.id}?alt=media`
       });
     } catch (err) {
       if (err.code === 404) {
@@ -166,14 +145,14 @@ function createDrive({
   }
   
   async function post(file, data) {
-    const body = new _FormData;
+    const body = new FormData;
     const meta = {
       name: file,
       parents: ["appDataFolder"]
     };
-    body.append("metadata", new _Blob([JSON.stringify(meta)], {type: "application/json; charset=UTF-8"}));
-    body.append("media", new _Blob([data], {type: "text/plain"}));
-    const result = await query({
+    body.append("metadata", new Blob([JSON.stringify(meta)], {type: "application/json; charset=UTF-8"}));
+    body.append("media", new Blob([data], {type: "text/plain"}));
+    const result = await request({
       method: "POST",
       path: "https://www.googleapis.com/upload/drive/v3/files?uploadType=multipart&fields=id,name,headRevisionId",
       body
@@ -187,10 +166,9 @@ function createDrive({
       return;
     }
     try {
-      await query({
+      await request({
         method: "DELETE",
-        path: `https://www.googleapis.com/drive/v3/files/${meta.id}`,
-        format: null
+        path: `https://www.googleapis.com/drive/v3/files/${meta.id}`
       });
     } catch (err) {
       if (err.code === 404) {
